refactor:VideoQoS
